Fixed issue with chaperone border renderer
Fixed issue where chaperone border renderer would be rotated incorrectly
if the camera rig (origin transform) had a rotation other than 0. This
is necessary for situations where the chaperone has to match the
orientation of elements of the world, but it is impossible to rotate the
whole world (e.g. Unity terrain)